relaciones hasta legajos ok

+++ src/main/java/com/guardias/backend/repositorio/TipoGuardiaRepositorio.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.guardias.backend.modelo.TipoGuardia;
 
 @Repository
-public interface TipoGuardiaRepositorio extends JpaRepository<TipoGuardia, Integer> {
+public interface TipoGuardiaRepositorio extends JpaRepository<TipoGuardia, Long> {
 
     Optional<TipoGuardia> findByNombre(String nombre);
     boolean existsByNombre(String nombre);
